mgr/cephadm: move NodeAssignment to new module
